feat(host): retarget core handlers to Gameplay.Api facades
Adapt Give/Clear, packet handlers, and SharedAssemblies to IPlayer/IItemStack/Api BlockPos and ISR wire handles.

+++ src/Orion/Entity/Traits/EntityAirSupplyTrait.cs
@@ -79,7 +79,7 @@ public sealed class EntityAirSupplyTrait : EntityTrait
             Entity,
             0.5f,
             null,
-            Entity.IsSwimming ? ActorDamageCause.Drowning : ActorDamageCause.Suffocation);
+            (int)(Entity.IsSwimming ? ActorDamageCause.Drowning : ActorDamageCause.Suffocation));
     }
 
     public int GetAirSupplyTicks()
